style: improve header
Refactor and differentiate headers from home page and logged in pages according to mockups.
It favors the use of classes as `nav` and `header` can be used in other contexts and we don't
want their styles to be applied everywhere.
This commit also introduces a `.container` class that will help limiting content width.