Fix checkbox alignment in Field forms
Use erb instead of ruby to concatenate strings. Hello whitespace-based alignment!

(Long story short: using <%= %> adds/keeps the whitespace around the ruby block, and since this is a text block, this results in something that is relevant in html. I could dig deeper, but since this erb pattern is common, let’s juste rely on it looking good)