Explicitly includes territory in layer rows
Since !381, we include the layer association fields. Before that, we eager-loaded all the object at the map level, in maps/show. This means we perform a specific query for each layer, and the eager-loading has to be done for each layer.

(We do N+1 queries, with N being the number of layers, which we can say is OK.)