Ensure rows are connected to their rowController
fixes #262

The connection order of RowController / MapController to their respective element is not guaranteed. However, in MapController#rowTargetConnected, we need the row.rowController to already be set, which only happens when the RowController itself is connected to its row element.

(We could expect controllers to be all connected to their _element_ before _targets_ are connected, but this is not the case anyway.)

--

I am not at ease with this fix: this makes the row_controller, the map_controller and map_stats even more interdependent. I’m starting to think we should connect the rows to the map manually in JS, without using stimulus.

--

Another smaller issue is that turbo preview (https://turbo.hotwired.dev/handbook/building#understanding-caching) restores the dom from cache, with the `data-map-target="row"` attribute already in place. This means the `mapController` (instantiated itself from the preview html) may be notifier briefly of the (cached) row.

There are already a number of issues with this. Maybe we just want to disable turbo caching globally with `<meta name="turbo-cache-control" content="no-cache">`.