Make sure to load the row geometry columns
followup !496

Almost everything works without it because we add a geojson attribute.
When updating a row field value, we also submit the geojson, which adds the correct geometry column to the record.

However, when changing a file field, we don’t touch the row geometry column, which breaks when validating (in take_first_of_geometry_collection) because self.geometry tries to use the point/line_string/polygon attribute, which does not exist.

refs #302