Tweak and note for conversion of files type value
This worksaround a failure when importing data which type_files columns. In a csv, type_files column get exported as strings; we want to ignore them, but now we are crashing (because value.map can't work if value is an empty string.)

This line:
`existing_blob_ids = values[field_id]`

is clumsy: we should pass the existing blob ids as params to the frontend and use the whole array of existing blob_ids (strings) and new files as a whole value.
This would help move this code to Field#cast.
Additionally, something like this is needed to be able to remove files.