Workaround a geojson territory import crash
This is actually due to a change in (I think) the latest release of libgeos. I suspect this is related to this: https://github.com/libgeos/geos/pull/671

On my machine, the TerritoryTest unit tests have started crashing after a brew upgrade; the bug happens somewhere in the rgeo wkt parsing, because some WKT string value has a Z marker and the expected value does not want a Z.

Apparently, the new behaviour of RGeo with the updated libgeos is to generate geometry with Z, even if not asked. Afterwards, if we try to parse this geometry (as text) to put it into our columns, rgeo will complain that the Z token is not expected.

This is solved by *not* converting the geometry to text. It’s a RGeo::Geos::CAPIPointImpl, and it does not have to be parsed before being inserted into PG.

I think there is something troubling underneath. Either the behaviour in the updated libgeos is buggy (and returns z values even when asked not to) or RGeo always had an issue that only showed up with the new behaviour.

Differences in behaviour:
On production
```
irb(main):001:0> RGEO_FACTORY.point(1, 1)
=> #<RGeo::Geos::CAPIPointImpl:0xbbbc "POINT Z (1 1 0)">
```

On my machine with libgeos 3.12.0
```
irb(main):002:0> RGEO_FACTORY.point(1, 1).as_text
=> "POINT Z (1 1 0)"
```