Allow serving the app on `hostname` in development
This makes debugging the app on a local network easier. Sometimes (especially when using roaming) the local wifi is ipv6 only, which makes typing the ip address on a phone inconvenient.

This means I can start the server on my mac go to `https://neon.local:3000` from my phone and it just works.